requirements.txt
Django>=4.1.4,<4.2
pymysql>=1.0.2,<1.1
djangorestframework>=3.14.0,<3.15
drf-spectacular>=0.25.1,<0.26
mysql-connector-python